% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_na.R
\name{check_for_na}
\alias{check_for_na}
\title{Check perturbed table for missingness in tabulation variables}
\usage{
check_for_na(DT, cols)
}
\arguments{
\item{DT}{-- \code{data.table}
Perturbed frequency table}

\item{cols}{-- \verb{character vector}
Tabulation variables}
}
\value{
Warning message if any tabulation variable contain missing values
}
\description{
Check perturbed table for missingness in tabulation variables
}
