% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_ptable.R
\name{generate_ptable_10_5_rule}
\alias{generate_ptable_10_5_rule}
\title{Generate ptable (10-5 rule)}
\usage{
generate_ptable_10_5_rule(max_pcv = 750, ckey_range = 255)
}
\arguments{
\item{max_pcv}{Max value for pcv. Default is 750.}

\item{ckey_range}{The max range for cell keys. Default is 255.}
}
\value{
A data.table assigning a pvalue to each ckey and pcv combination
}
\description{
\code{generate_ptable_10_5_rule()} generates a sample p-table based on 10-5 rule,
which means a suppression threshold of 10 and rounding to the nearest 5.
}
\examples{
ptable <- generate_ptable_10_5_rule()

}
