% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{micro}
\alias{micro}
\title{Example data (micro)}
\format{
A data.table containing 1000 observations of 11 variables
}
\usage{
data(micro)
}
\description{
A data set containing randomly generated data to showcase the cell key
perturbation method.
}
\details{
\itemize{
\item record_key. record key value (0-255)
\item var1. example variable 1 (1-5)
\item var2. example variable 2 (1,2)
\item var3. example variable 3 (1-4)
\item var4. example variable 4 (1-4)
\item var5. example variable 5 (1-10)
\item var6. example variable 6 (1-5)
\item var7. example variable 7 (1-5)
\item var8. example variable 8 (A-D)
\item var9. example variable 9 (A-H)
\item var10. example variable 10 (1-49)
}
}
\keyword{datasets}
