% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptable_10_5}
\alias{ptable_10_5}
\title{Perturbation table}
\format{
A data.table containing 192000 observations of 3 variables
}
\usage{
data(ptable_10_5)
}
\description{
A data set containing the rules to apply cell key perturbation with a
threshold of 10, and rounding to base 5.
}
\details{
\itemize{
\item pcv. perturbation cell value (1-750)
\item ckey. cell key value
\item pvalue. perturbation value to be applied
}
}
\keyword{datasets}
