% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_perturbed_table.R
\name{create_perturbed_table}
\alias{create_perturbed_table}
\title{Create a frequency table which has has a cell key perturbation applied}
\usage{
create_perturbed_table(data, geog, tab_vars, record_key_arg, ptable)
}
\arguments{
\item{data}{A data.table containing the data to be tabulated and perturbed}

\item{geog}{A string vector giving the column name in 'data' that contains the desired geography level for the frequency table}

\item{tab_vars}{A string vector giving the column names in 'data' of the variables to be tabulated.}

\item{record_key_arg}{A String containing the column name in 'data' giving the record keys required for perturbation.}

\item{ptable}{A data.table containing the 'ptable' file which determines when perturbation is applied.}
}
\value{
Returns a data.table giving a frequency table which has had cell key perturbation applied according to the ptable supplied.
}
\description{
'create_perturbed_table()' creates a frequency table which has has a
cell key perturbation technique applied with help from a p-table which
determines how perturbation is applied.
The perturbation adds a small amount of noise to some cells in a table,
meaning that users cannot be sure whether differences between tables
represent a real person, or are caused by the perturbation.
Cell Key Perturbation is consistent and repeatable, so the same cells are
always perturbed in the same way.
}
\examples{
micro <- fread("census_2011.csv")
ptable_10_5 <- fread("ptable_10_5_rule.csv")

geog <- c("Region")
tab_vars <- c("Age","Health","Occupation")
record_key_arg <-"record_key"

perturbed_table  <- create_perturbed_table(micro, geog, tab_vars, record_key_arg, ptable_10_5)

Using direct inputs, and selecting no geography breakdown:

perturbed_table <-create_perturbed_table(data = micro,
                                         record_key = "Record_key",
                                         geog = c(),
                                         tab_vars = c("Sex","Industry","Occupation"),
                                         ptable = ptable_10_5)

}
