% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_perturbed_table.R
\name{create_perturbed_table}
\alias{create_perturbed_table}
\title{Create a frequency table with cell key perturbation applied}
\usage{
create_perturbed_table(data_table, geog, tab_vars, record_key_arg, ptable)
}
\arguments{
\item{data_table}{A data.table containing the data to be tabulated and perturbed}

\item{geog}{A string vector giving the column name in 'data' that contains the desired geography level for the frequency table}

\item{tab_vars}{A string vector giving the column names in 'data' of the variables to be tabulated.}

\item{record_key_arg}{A String containing the column name in 'data' giving the record keys required for perturbation.}

\item{ptable}{A data.table containing the 'ptable' file which determines when perturbation is applied.}
}
\value{
Returns a data.table giving a frequency table which has had cell key perturbation applied according to the ptable supplied.
}
\description{
'create_perturbed_table()' creates a frequency table which has had
cell key perturbation applied to the counts.
A p-table file needs to be supplied which determines which cells are perturbed.
The data needs to contain a 'record key' variable which along with the
ptable allows the process to be repeatable and consistent.
}
\examples{
geog <- c("var1")
tab_vars <- c("var5","var8")
record_key_arg <-"record_key"

perturbed_table  <- create_perturbed_table(micro, geog, tab_vars, record_key_arg, ptable_10_5)

perturbed_table <-create_perturbed_table(data_table = micro,
                                         record_key = "record_key",
                                         geog = c(),
                                         tab_vars = c("var1","var5","var8"),
                                         ptable = ptable_10_5)

}
