% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate sample microdata}
\usage{
generate_test_data(size, rkey_range = 255)
}
\arguments{
\item{size}{Number of rows in the sample microdata}

\item{rkey_range}{The max range for record keys. Default is 255.}
}
\value{
A data.table containing randomly generated microdata and record keys
}
\description{
\code{generate_test_data()} creates a sample microdata containing randomly
generated microdata columns and record keys for testing purposes.

Note: A seed is set for random value generator to obtain same output in
different runs. However, the sample microdata included in the package will
be different than this one, as it was generated from the corresponding
python package for consistency in test output.
}
\examples{
data <- generate_test_data(size = 1000)
data <- generate_test_data(size = 1000, rkey_range = 255)
}
